# docs/swag/wallet/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Wallet API server",
        "title": "Wallet API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "ap11.0dev.cc",
    "basePath": "/",
    "paths": {
...
        "/wallet_transactions/cash_flow": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a filtered and paginated list of wallet transactions within a date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet Transactions"
                ],
                "summary": "Get transaction cash flow",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Start date for filtering (YYYY-MM-DD format)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "End date for filtering (YYYY-MM-DD format)",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transaction list",
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.GetListTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid date format",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
        "/wallet_transactions/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new transaction for the specified wallet with transaction details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet Transactions"
                ],
                "summary": "Create wallet transaction",
                "parameters": [
                    {
                        "description": "Transaction creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.CreateTransactionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created transaction",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid transaction data",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apiresp.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errCode": {
                    "type": "integer"
                },
                "errDlt": {
                    "type": "string"
                },
                "errMsg": {
                    "type": "string"
                }
            }
        },
...
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.CreateTransactionReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "createdBy": {
                    "type": "string"
                },
                "descriptionEn": {
                    "type": "string"
                },
                "descriptionZh": {
                    "type": "string"
                },
                "entryType": {
                    "type": "string"
                },
                "impactedItem": {
                    "type": "integer"
                },
                "referenceCode": {
                    "type": "string"
                },
                "transactionType": {
                    "type": "string"
                },
                "walletID": {
                    "type": "integer"
                }
            }
        },
...
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.GetListTransactionResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.WalletTransaction"
                    }
                }
            }
        },
...
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.WalletTransaction": {
            "type": "object",
            "properties": {
                "afterBalance": {
                    "type": "number"
                },
                "amount": {
                    "type": "number"
                },
                "beforeBalance": {
                    "type": "number"
                },
                "descriptionEn": {
                    "type": "string"
                },
                "descriptionZh": {
                    "type": "string"
                },
                "entryType": {
                    "type": "string"
                },
                "impactedItem": {
                    "type": "integer"
                },
                "isShown": {
                    "type": "boolean"
                },
                "referenceCode": {
                    "type": "string"
                },
                "transactionDate": {
                    "type": "string"
                },
                "transactionType": {
                    "type": "string"
                },
                "walletID": {
                    "type": "integer"
                },
                "walletTransactionID": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}